back/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.11
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.14.0
Flask==3.1.0
Flask-Mail==0.10.0
fonttools==4.55.0
fuzzywuzzy==0.18.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
Levenshtein==0.26.1
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python3-openid==3.2.0
pytz==2024.1
RapidFuzz==3.10.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.1.3
xmltodict==0.14.2
